jasmine productsspec

+++ migrations/sqls/20230124084607-products-table-up.sql
@@ -1,2 +1,2 @@
 /* Replace with your SQL commands */
-CREATE TABLE products (id SERIAL PRIMARY KEY, name VARCHAR(255), price bigint, category VARCHAR(255));
+CREATE TABLE products (id SERIAL PRIMARY KEY, name VARCHAR(255), price integer, category VARCHAR(255));
